+++ dlpdes/model/mlp.py
@@ -17,9 +17,9 @@ class MLP(nn.Module):
     """
     def __init__(self, args):
         super().__init__()
-        self.depth = getattr(args, "mlp_depth", 5)
+        self.depth = getattr(args, "mlp_depth", 1)
         input_size = getattr(args, "input_size", 2)
-        hidden_size = getattr(args, "mlp_hidden_size", 30)
+        hidden_size = getattr(args, "mlp_hidden_size", 200)
         output_size = getattr(args, "output_size", 1)
 
         self.activation = nn.Tanh()