+++ dlpdes/ps_run.py
@@ -14,42 +14,33 @@ def parse_args():
     
     # --- basic config ---
     parser.add_argument("--eq", type=str, default="poisson", help="Equation name (poisson, ac, etc.)")
-    parser.add_argument("--device", type=str, default="cuda:3")
+    parser.add_argument("--device", type=str, default="cuda:4")
     parser.add_argument("--model", type=str, default="mlp", help="Model architecture (moe_d, mlp , moe_d_w etc.)")
     parser.add_argument("--seed", type=int, default=2026)
     
     # --- sampling ---
     parser.add_argument("--sample_method", type=str, default="grid") # random or grid
-    parser.add_argument("--Nf", type=int, default=10000)
-    parser.add_argument("--Nb", type=int, default=200)
-    parser.add_argument("--nx", type=int, default=10)
-    parser.add_argument("--ny", type=int, default=10)
-    parser.add_argument("--n_per_edge", type=int, default=10)
+    parser.add_argument("--Nf", type=int, default=5000)
+    parser.add_argument("--Nb", type=int, default=100)
+    parser.add_argument("--nx", type=int, default=30)
+    parser.add_argument("--ny", type=int, default=30)
+    parser.add_argument("--n_per_edge", type=int, default=100)
     parser.add_argument("--w_pde", type=float, default=1.0)
     parser.add_argument("--w_bc", type=float, default=1.0)
 
 
-    # --- modeling ---
-    parser.add_argument("--input_size", type=int, default=2)
-    parser.add_argument("--output_size", type=int, default=1)
-    parser.add_argument("--num_experts", type=int, default=3)
-    parser.add_argument("--hidden_size", type=int, default=20)
-    parser.add_argument("--depth", type=int, default=2)
-    parser.add_argument("--gating_hidden_size", type=int, default=5)
-    parser.add_argument("--gating_depth", type=int, default=1)
+    # --- modeling ---s
     parser.add_argument("--use_double", action="store_true", default=True)
     
     # --- training ---
-    parser.add_argument("--iters", type=int, default=5000)  # adam iters
+    parser.add_argument("--lm_epochs", type=int, default=10000)  # lm epochs
     parser.add_argument("--lm_beta_train", type=bool, default=False) # lbfgs lr
     parser.add_argument("--lbfgs_iter", type=int, default=700) # lbfgs iters
     parser.add_argument("--lbfgs_lr", type=float, default=1.0) # lbfgs lr
     parser.add_argument("--lr", type=float, default=8e-4) # adam lr
-    parser.add_argument("--use_scheduler", action="store_true", default=False)
     parser.add_argument("--lr_step_size", type=int, default=2000)
     parser.add_argument("--lr_gamma", type=float, default=0.8) 
-    parser.add_argument("--log_freq", type=dict, default={"adam": 500, "lbfgs": 10,"proj_adam":500,"lm":100}) # in trainer.py for print loss
-    parser.add_argument("--proj_g_update_freq", type=int, default=100) # in trainer.py for updating projection g
+    parser.add_argument("--log_freq", type=dict, default={"adam": 500, "lbfgs": 10,"proj_adam":500,"lm":20}) # in trainer.py for print loss
     parser.add_argument("--checkpoint_freq", type=dict, default={"adam": 10000, "lbfgs": 200,"proj_adam":10000,"lm":10000}) # in checkpoint_callback.py for saving model
 
     
@@ -96,6 +87,8 @@ def main():
     # # pipe.trainer.train_lbfgs(pipe.data)
     # pipe.trainer.train_proj_adam(pipe.data)
     pipe.trainer.train_lm(pipe.data)
+    # pipe.refresh_data()
+    # pipe.trainer.train_lm(pipe.data)
     print(f"--- {args.eq.upper()} train finished ---")